Analysis on "mean" predictions
The MM prediction is synthesized from mostly means of the predictions

Then, who is further away from the means on average?

The two folks in the middle of the pack, SU and HR, have very different
percentage-paths-to-victory. What's up with that?  SU is the guesser with the
the most "normal" guesses out of all of us! His guesses average only 3.52
"slots" away from the "Mean-Median" guess-ordering I created. So it'll be
really hard for him to distinguish himself from the pack. He's really banking
on N and T at this point; everything else is solidly middling but not
distinguishing. On the flip-side, he's tended to predict with "the wisdom of
the crowds", so he'll probably do pretty well on average.

HR is the guesser with the least "normal" guesses out of all of us! her guesses
average 6.19 "slots" away from MM. There's lots of room for things to break her
way and she becomes the clear leader. B, K, R, and S coming true would all be
big swings towards her, and thus she's got lots of paths to victory. The caveat
is that other folks had different predictions on average, so against "the
wisdom of the crowds"

The Mean-Average-Error for each of us, compared against the "MM" prediction I
deduced from our collective means and medians:
{'HR': 6.19, 'LE': 5.9, 'AA': 5.81, 'NT': 5.81, 'CJ': 5.81, 'BF': 5.62, 'DJ':
5.52, 'NA': 5.33, 'EK': 5.24, 'PJ': 5.05, 'PG': 4.86, 'JC': 4.86, 'AL': 4.86,
'TR': 4.76, 'MA': 4.48, 'JL': 4.48, 'JJ': 4, 'SU': 3.52}